refactor: teamsyearcomponent is nu daadwerkelijk een component, en teamscomponent zit nu onder molecules ipv organisms
