+++ src/lib/components/organisms/TeamsYearComponent.svelte
@@ -1,17 +1,38 @@
-<!-- <script>
-    // // TODO: actually implement this component
-	import TeamsComponent from '$lib/components/organisms/TeamsComponent.svelte'
-    let { year } = $props();
-    // const startDate = new Date(teams.time_start);
-    // const endDate = new Date(teams.time_end);
+<script>
+	/* eslint-disable svelte/require-each-key */
+	import TeamsComponent from '$lib/components/molecules/TeamsComponent.svelte'
 
-    // const month = ["januari","februari","maart","april","mei","juni","juli","augustus","september","october","november","december"];
-    console.log("tem" + teams.start_date);
+	let { year } = $props()
 </script>
 
+<article>
+	<h3>{year[0]}</h3>
+	<ul>
+		{#each year[1] as teams}
+			<TeamsComponent {teams} />
+		{/each}
+	</ul>
+</article>
 
-    <article>
-			{#each year as teams(year.id)}
-				<TeamsComponent teams={teams}/>
-			{/each}
-		</article> -->
+<style>
+	h3 {
+		font-size: 1.5rem;
+		padding-block: 1em;
+	}
+
+	article {
+		display: flex;
+		flex-direction: column;
+	}
+
+	ul {
+		list-style: none;
+		display: flex;
+		flex-direction: column;
+
+		@media (min-width: 450px) {
+			gap: 5vw;
+			flex-direction: row;
+		}
+	}
+</style>
